Renaming api/errors/etcd to api/errors/storage as it no longer has any etcd specific dependencies.  Reference issue #17546
